Added select stmt assembly to query
