--- src/sql_orm/select/help.js
@@ -1,22 +1,0 @@
-/*
-For help query.
-
-Query to get columns and types and comments.
-
-Return information from pages array
-
-
-*/
-
-//Parse json object.
-
-//pull default json object
-//search postgres
-//
-
-//table/view/function comments
-//column comments
-//column types
-//grid table params
-//permissions
-//contact for help
