+++ src/sql_orm/select/where/search.js
@@ -1,5 +1,6 @@
 /*
 
+Not implemented
 //search routes.
 
 return rows
